Boton volver

+++ src/app/_components/cargar-receta-completa/cargar-receta-completa.component.ts
@@ -223,7 +223,7 @@ export class CargarRecetaCompletaComponent implements OnInit {
 
   cancelar(volver: boolean) {
     if(!volver) {
-      if (this.errorMensaje == null) {
+      if (this.errorMensaje == null && this._imageFile != undefined) {
         this._RecetaService.deleteImage(this._imageName).subscribe();
       }
       this.finalizarCarga.emit(false);
